backend/auth/requirements.txt
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
psycopg2-binary==2.9.9